+++ src/main/scala/algorithms/parallel/BFS.scala
@@ -6,14 +6,14 @@ package algorithms.parallel
  */
 import graph.{ Edge, EdgeProvider, SimpleEdge }
 
-class BFS(root: Long)(implicit eps: Seq[EdgeProvider[SimpleEdge]]) extends Algorithm[Long] {
+class BFS(root: Int)(implicit eps: Seq[EdgeProvider[SimpleEdge]]) extends Algorithm[Int] {
   def iterations() = {
-    var level = 1L
+    var level = 1
     scatter(root, level)
     update
 
     while (!gather.isEmpty) {
-      level += 1L
+      level += 1
       for (ep <- eps.par; Edge(u, v) <- ep.getEdges if (gather(u) && !data.contains(v))) {
         scatter(v, level)
       }
@@ -22,14 +22,14 @@ class BFS(root: Long)(implicit eps: Seq[EdgeProvider[SimpleEdge]]) extends Algor
   }
 }
 
-class BFS_U(root: Long)(implicit eps: Seq[EdgeProvider[SimpleEdge]]) extends Algorithm[Long] {
+class BFS_U(root: Int)(implicit eps: Seq[EdgeProvider[SimpleEdge]]) extends Algorithm[Int] {
   def iterations() = {
-    var level = 1L
+    var level = 1
     scatter(root, level)
     update
 
     while (!gather.isEmpty) {
-      level += 1L
+      level += 1
       for (ep <- eps.par; Edge(u, v) <- ep.getEdges) {
         if (gather(u) && !data.contains(v)) scatter(v, level)
         if (gather(v) && !data.contains(u)) scatter(u, level)
